chore(component): consolidate google map component
